feat: 회원가입 시 숏링크 제작되도록 함
