+++ src/main/java/com/woozuda/backend/shortlink/Service/ShareService.java
@@ -101,57 +101,6 @@ public class ShareService {
 
     }
 
-    public List<String> extractContent(List<NoteContent> notecontents){
-
-        return null;
-    }
-
-    @Transactional
-    public ShortLinkDto makeShortLink(String username){
-
-        UserEntity userEntity = userRepository.findByUsername(username);
-
-        //이미 해당 계정에는 숏링크가 존재함 . 예외처리
-        if(shortLinkRepository.findByUserEntity(userEntity) != null){
-            return null;
-        }
-
-        String newShortLink = "";
-
-        while(true){
-
-            // 숏링크 생성
-            newShortLink = createRandomLink();
-
-            // 중복 검사 (다른 유저랑 동등한 숏링크 인지?)
-            if(shortLinkRepository.findByUrl(newShortLink) == null){
-                break;
-            }
-        }
-
-        ShortLink shortLink = new ShortLink(null, newShortLink, userEntity);
-
-        shortLinkRepository.save(shortLink);
-
-        return new ShortLinkDto(newShortLink);
-    }
-
-    public String createRandomLink(){
-
-        String str = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
-        int strLength = str.length();
-
-        Random random = new Random();
-
-        StringBuffer randomStr = new StringBuffer();
-
-        for (int i = 0; i < 8; i++) {
-            randomStr.append(str.charAt(random.nextInt(strLength)));
-        }
-
-        return randomStr.toString();
-    }
-
     @Transactional
     public ShortLinkDto getShortLink(String username) {
